refactor NavBar with splitting Desktop/Mobile NavBar
