+++ src/utils/CustomSwal.ts
@@ -1,0 +1,9 @@
+import Swal from 'sweetalert2';
+
+export const Toast = Swal.mixin({
+    toast: true,
+    position: 'top',
+    showConfirmButton: false,
+    timer: 1500,
+    timerProgressBar: true,
+  });
